Remove dead code
