+++ src/gurobi_optimods/opf/__init__.py
@@ -1,8 +1,4 @@
 from gurobi_optimods.opf.api import (  # noqa: F401
     compute_violations_from_given_voltages,
-    read_case_from_mat_file,
-    read_coords_from_csv_file,
-    read_voltages_from_csv_file,
     solve_opf_model,
-    turn_result_into_mat_file,
 )
