+++ src/gurobi_optimods/opf/grbcasereader.py
@@ -1,10 +1,6 @@
 import cmath
 import logging
 import math
-import time
-
-import numpy as np
-import scipy
 
 logger = logging.getLogger(__name__)
 
@@ -645,257 +641,3 @@ def convert_case_to_internal_format(case_dict):
         )
 
     return alldata
-
-
-def read_case_file_mat(casefile):
-    """
-    Reads case data from a `.mat` file following MATPOWER notation and
-    returns an OptiMod compatible dictionary holding all relevant case data
-
-    :param casefile: Path to `.mat` file holding all case relevant information
-    :type casefile: str
-
-    :raises ValueError: Missing field
-
-    :return: A dictionary holding all case relevant data
-    :rtype: dict
-    """
-
-    starttime = time.time()
-    case_dict = {}
-
-    logger.info("Reading case file %s." % casefile)
-    mat = loadmat(casefile)
-
-    if "mpc" not in mat.keys():
-        raise ValueError("Provided .mat file does not have an mpc field")
-
-    mpc = mat["mpc"]
-
-    for x in ["baseMVA", "bus", "gen", "branch", "gencost"]:
-        if x not in mpc.keys():
-            raise ValueError(f"Provided .mat file does not have a {x} field")
-
-    case_dict["baseMVA"] = mpc["baseMVA"]
-    mpcbuses = mpc["bus"]
-    mpcgen = mpc["gen"]
-    mpcbranch = mpc["branch"]
-    mpcgencost = mpc["gencost"]
-
-    # Buses
-    numbuses = 0
-    buses = {}
-    if mpcbuses.ndim == 1:
-        raise ValueError("Provided .mat files has only 1 bus")
-    for b in mpcbuses:
-        numbuses += 1
-        buses[numbuses] = {}
-        buses[numbuses]["bus_i"] = int(b[0])
-        buses[numbuses]["type"] = int(b[1])
-        buses[numbuses]["Pd"] = b[2]
-        buses[numbuses]["Qd"] = b[3]
-        buses[numbuses]["Gs"] = b[4]
-        buses[numbuses]["Bs"] = b[5]
-        buses[numbuses]["area"] = b[6]
-        buses[numbuses]["Vm"] = b[7]
-        buses[numbuses]["Va"] = b[8]
-        buses[numbuses]["baseKV"] = b[9]
-        buses[numbuses]["zone"] = b[10]
-        buses[numbuses]["Vmax"] = b[11]
-        buses[numbuses]["Vmin"] = b[12]
-        if buses[numbuses]["type"] == 3:
-            slackbus = buses[numbuses]["bus_i"]
-            logger.info(f"    Slack bus: {slackbus}")
-            logger.info(
-                f"    Bus {numbuses} ID {buses[numbuses]['bus_i']} is the reference bus."
-            )
-
-    case_dict["bus"] = buses
-
-    # Generators
-    numgens = 0
-    gens = {}
-    if mpcgen.ndim == 1:
-        mpcgen = [mpcgen]
-    for g in mpcgen:
-        numgens += 1
-        gens[numgens] = {}
-        gens[numgens]["bus"] = int(g[0])
-        gens[numgens]["Pg"] = g[1]
-        gens[numgens]["Qg"] = g[2]
-        gens[numgens]["Qmax"] = g[3]
-        gens[numgens]["Qmin"] = g[4]
-        gens[numgens]["Vg"] = g[5]
-        gens[numgens]["mBase"] = g[6]
-        gens[numgens]["status"] = 0 if g[7] <= 0 else 1
-        gens[numgens]["Pmax"] = g[8]
-        gens[numgens]["Pmin"] = g[9]
-        gens[numgens]["Pc1"] = g[10]
-        gens[numgens]["Pc2"] = g[11]
-        gens[numgens]["Qc1min"] = g[12]
-        gens[numgens]["Qc1max"] = g[13]
-        gens[numgens]["Qc2min"] = g[14]
-        gens[numgens]["Qc2max"] = g[15]
-        gens[numgens]["ramp_agc"] = g[16]
-        gens[numgens]["ramp_10"] = g[17]
-        gens[numgens]["ramp_30"] = g[18]
-        gens[numgens]["ramp_q"] = g[19]
-        gens[numgens]["apf"] = g[20]
-
-    case_dict["gen"] = gens
-
-    # Branches
-    numbranches = 0
-    branches = {}
-    if mpcbranch.ndim == 1:
-        mpcbranch = [mpcbranch]
-    for b in mpcbranch:
-        numbranches += 1
-        branches[numbranches] = {}
-        branches[numbranches]["fbus"] = b[0]
-        branches[numbranches]["tbus"] = b[1]
-        branches[numbranches]["r"] = b[2]
-        branches[numbranches]["x"] = b[3]
-        branches[numbranches]["b"] = b[4]
-        branches[numbranches]["rateA"] = b[5]
-        branches[numbranches]["rateB"] = b[6]
-        branches[numbranches]["rateC"] = b[7]
-        branches[numbranches]["ratio"] = 1.0 if b[8] == 0.0 else b[8]
-        branches[numbranches]["angle"] = b[9]
-        branches[numbranches]["status"] = b[10]
-        branches[numbranches]["angmin"] = b[11]
-        branches[numbranches]["angmax"] = b[12]
-
-    case_dict["branch"] = branches
-
-    # Generator costs
-    numgencosts = 0
-    gencosts = {}
-    if mpcgencost.ndim == 1:
-        mpcgencost = [mpcgencost]
-    for g in mpcgencost:
-        numgencosts += 1
-        gencosts[numgencosts] = {}
-        gencosts[numgencosts]["costtype"] = g[0]
-        gencosts[numgencosts]["startup"] = g[1]
-        gencosts[numgencosts]["shutdown"] = g[2]
-        degree = int(g[3])
-        gencosts[numgencosts]["n"] = degree
-        costvector = [g[j] for j in range(4, 4 + degree)]
-        gencosts[numgencosts]["costvector"] = costvector
-
-    if numgencosts > numgens:
-        # FIXME: spec says we can have twice as many entries, representing reactive power.
-        # Do we not handle this case?
-        raise ValueError(f"Read {numgencosts} gen costs but only {numgens} generators.")
-
-    case_dict["gencost"] = gencosts
-
-    endtime = time.time()
-    logger.info(f"Reading and building time: {endtime - starttime} s.")
-
-    return case_dict
-
-
-def loadmat(filename):
-    """
-    This function should be called instead of direct scipy.io.loadmat
-    as it cures the problem of not properly recovering Python dictionaries
-    from mat files. It calls the function check keys to cure all entries
-    which are still mat-objects
-
-    :param filename: Path to `.mat` file holding all case relevant information
-    :type filename: str
-
-    :return: Dictionary holding all case relevant data read from the given `.mat` file
-    :rtype: dict
-    """
-    data = scipy.io.loadmat(filename, struct_as_record=False, squeeze_me=True)
-    return _check_keys(data)
-
-
-def _check_keys(dict):
-    """
-    Checks if entries in dictionary are mat-objects. If yes
-    todict is called to change them to nested dictionaries
-
-    :param dict: Dictionary holding all case data read from a `.mat` file by the
-                 scipy.io.loadmat function
-    :type dict: dict
-
-    :return: Dictionary holding all case relevant data read from the given `.mat` file
-    :rtype: dict
-    """
-    for key in dict:
-        if isinstance(dict[key], scipy.io.matlab.mat_struct):
-            dict[key] = _todict(dict[key])
-    return dict
-
-
-def _todict(matobj):
-    """
-    A recursive function which constructs from matobjects nested dictionaries
-
-    :param matobj: Placeholder for holding read data from structs
-    :type matobj: scipy.io.matlab.mat_struct
-
-    :rtype: dict
-    :return: A possibly nested dictionary with all case relevant data
-    """
-    dict = {}
-    for strg in matobj._fieldnames:
-        elem = matobj.__dict__[strg]
-        if isinstance(elem, scipy.io.matlab.mat_struct):
-            dict[strg] = _todict(elem)
-        else:
-            dict[strg] = elem
-    return dict
-
-
-def turn_opf_dict_into_mat_file(solution, filename):
-    """
-    Writes a `.mat` file out of an OPF solution dictionary
-
-    :param solution: OPF solution dictionary
-    :type solution: dict
-    :param filename: Name of `.mat` file where to write the solution data
-    :type filename: str
-    """
-
-    # Buses
-    buses = solution["bus"]
-    matrix = []
-    for bus in buses.values():
-        matrix.append(list(bus.values()))
-
-    solution["bus"] = np.array(matrix)
-    # Generators
-    gens = solution["gen"]
-    matrix = []
-    for gen in gens.values():
-        matrix.append(list(gen.values()))
-
-    solution["gen"] = np.array(matrix)
-    # Branches
-    branches = solution["branch"]
-    matrix = []
-    for branch in branches.values():
-        matrix.append(list(branch.values()))
-
-    solution["branch"] = np.array(matrix)
-    # Generator costs
-    gencosts = solution["gencost"]
-    matrix = []
-    # For generator costs, the last dictionary entry is a list of values
-    for genc in gencosts.values():
-        l = list(genc.values())
-        costvector = l[-1]
-        l.pop()
-        for item in costvector:
-            l.append(item)
-
-        matrix.append(l)
-
-    solution["gencost"] = np.array(matrix)
-    # Write mat file
-    scipy.io.savemat(filename, {"result": solution})
